+++ src/components/Modal/Dialog/index.ts
@@ -1,3 +1,4 @@
+export * from './Confirm'
 export * from './Dialog'
 export * from './DialogManager'
 export * from './DialogSurface'